LayoutTests/fast/preloader/image-srcset.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpResourceResponseMIMETypes();
    }

    addEventListener("load", function() {
        if (internals)
            shouldBeFalse('internals.isPreloaded("resources/image1.png")');
    }, false);
</script>
<script src="resources/non-existant.js"></script>
</head>
<body>
<div>This test passes if the fallback image (defined in the src attribute) is not preloaded, but the 1x version (defined in srcset) is.</div>
<img height="100" width="100" src="resources/image1.png" srcset="resources/base-image1.png 1x, resources/base-image3.png 3x, resources/base-image2.png 2x"></img>
</body>
</html>
